Fix the my meetings page
We have splitted the single meeting from the recursive meeting. This allows to print only the next
recursive meeting in the my meeting page (and not all of them).

In addition, update the index function to print an empty page if there is no calendar available in
the database (ie: fedocal has just been created)