Fix the library
For some reasons, I used d=d+8 to get the last day of the week instead
of day=day+6 which is the correct calculation.

I made similar error to get next week's date.

Fix some documentation errors

Check if there is a recursion before checking if there is a recursion
frequency or a recursion end date