Fix the test_fedocallib add_meeting tests
It raises an InvalidMeeting error not an AttributeError but only
when the date is enough in the past to not be in the future.

TODAY refers to the Tuesday of the current week, if the tests are
ran on a Monday, then TODAY is actually in the future and the tests
does not raise a InvalidMeeting but an AttributeError in this case
(pretty much like the first test in this method).

By changing the date to TODAY - 4 days, we are sure that we are
trying to add a meeting in the past (last week), so the output
of the test should be always consistent.