Move the API from the main controller to its own.
The controller of the API is moved to its own file instead of being in the main controller.

Fix the tests accordingly, this meant fixing the test configuration file as well as the
__init__ method.

Adding as well a new meeting (#11) in order to keep the coverage of the API at 100%.

In the meanwhile, change the setUp function for all tests to use scoped_session as in
the application.