When editing or adding a recursive meeting check that the time slot is always available
This is aiming at preventing meeting conflicting later in the recursion.

Additionally, when editing a recursive meeting we keep the old one as long as there the
meeting used to be part of a recursion, we don't need to know if the new meeting is
part of a recursion or not.

Fix the unit-tests accordingly to keep our coverage.

Fixes ticket 18