Add a list view
In the main menu there is now an entry to see all the meetings of the year, year/month or
even year/month/day for a given calendar.
Update the documentation accordingly to reflect these possibilities.
This commit adds a new dependency to python-dateutil (version 1.5 or lower).

The recursive meetings should be added as if they were individuals.