Back-end work
Rework the get_by_date function. It can be used in more places
giving simpler and more robust code.

Add function to expand recursive meetings, we need this at several
places so we need one central place to reuse it.

Adjust the unit-test for these changes, since we expand the recursive
meeting that does return more results in a number of places.