Add markdown syntax for the meeting description
This commit adds a new jinja decorator that converts the given
string into html using the Markdown syntax from the python-markdown
library.

This decorator is used on the information field of the meetings and
allows providing links, titles to the information.

The template macro render_field_in_row has been adjusted to allow
turning off the autoescape for the after field using the parameter
``escape_after``.