Fix queries
Apparently SQLAlchemy does not like queries with 'is not None' so
they are changed back to '!= None'.

There was a bug in get_active_regular_meeting which was checking
if the end of the regular meeting is after the end of the week, this
leading to the meeting not showing up in its last week.