Move the API documentation inside the API code itself and adjust it to return real json
Moving the API documentation inside the API allows to use it in the sphinx
documentation as well as to generate the online documentation.

So far we were returning a fake json generated by hand, do the things correctly now and
use the json module for that.

Return the arguments provided by the user with the results.

Use HTTP status code to indicate when something went wrong.

Adjust the unit-tests to reflect all this.