Implement the Meeting.get_overlaping_meetings
This method retrieves all the meeting which are occuring in the viscinity of the date
provided. It will return all the meetings surrounding the dates, the meetings included
in these dates and the meeting surrounding either the start or stop provided dates.

This covers the use-case of full-day meeting as well as multiple-days meetings.