Fix the cla_plus_one_required decorator
The redirection to auth_login is a bad idea has if the user is already logged in there,
fas-openid will redirect to fedocal which will redirect to fas-openid, creating an infinite
loop of redirection.

This commit should fix it.

More discussion about this point at: https://github.com/fedora-infra/nuancier-lite/issues/11