Drop the code to check if the agenda is free at this date/time
As the dates and times are not stored in the database using their specified timezone,
we can no longer rely on a DB query to check is the agenda is free or not, it would need
to convert all the meetings into the same timezone to be able to make the check
consistent.

The code was doing it with get_by_date but that was not taking into account dates with
timezones where the day itself would be changed.

Drop the unit-tests with the code