Add a wiki_link field when adding a meeting.
This field is then added to the description of the meeting. This might be redundant but
emphasis to the user the fact that he/she might have to fill in a wiki page.

We check that the link isn't present in the description before adding it.

Fixes https://fedorahosted.org/fedocal/ticket/77