Fix the timezone names when loading an iCal file
In iCal timezone names are most often of the type EDT or CEST or IST, but
IST can be applied to the zones:
   ['Asia/Colombo', 'Asia/Kolkata', 'Europe/Dublin']
So we need to keep a mapping of the localized timezone name and their UTC
offset to be able to track down which is which.
We also define some standard timezone acronym (EDT, CET, CEST) for which
we can directly substitute the name for one pytz and datetime know.

This should allow us to load our own iCal files.