Fix the links and javascript in the list view to fix deleting a specific meeting
To delete a specific meeting in a recursion we split the recursion into
two and drop the meeting in the middle.
But to do this, we need to know the from_date argument.

The agenda view had this working but for some reasons it was missing from
the list view.