Fix deleting a recursive meeting in the middle of the recursion
When deleting a recursive meeting, what we need to check is if the
recursion ends before the next occurrence of the meeting. That occurrence
is meeting_date + meeting_frequency. cnt would only have to do here if
meeting_date was the original meeting.meeting_date (which it is not).

Fixes https://fedorahosted.org/fedocal/ticket/135