When editing a meeting which has a date specified in the url, use that date
Currently it was going one meeting to far in the future because it was
looking at the next occurrence of the meeting from next week.
This is actually not the intended behavior. The function
update_date_rec_meeting is meant to do it and does it well, so let's rely
on it