Rather than disabling, hide the timezone field for full-day meeting
Full-day meetings are hard-coded to be from D 00:00 to D+X 00:00 UTC.
Until now we used to disable the timezone field, which meant it was not
sent when the form was submitting, leading to an 'invalid choice' error.
Now, we just set it to UTC and hide it.
This way it is sent as part of the form but it is not shown in the UI
which is what we want since the user cannot change the TZ for full-day
meetings.

Fixes https://fedorahosted.org/fedocal/ticket/150