Drop the `meeting` variable and adjust `next_meeting` to the timezone
We were using the next_meeting object that we did not update for timezone
information, so this lead to glitches such as the one reported at:
https://fedorahosted.org/fedocal/ticket/160
With this change, we update the next_meeting update for the timezone of
interest.

While doing so, we realized that the meeting object was basically
un-needed, we do not really use it and there is not much interest in
keeping it around. So we are removing it and relying on the orig_meeting
object (original meeting) when we need some information.