fix is_admin still looking at cla_done
previously, even if a user was in the admins group, and had signed the
FPCA, and were a member of more than one group (CLA+1), they were not
given access to the admin interface. This was due to the fact that the
is_admin function was still checking for cla_done rather than
signed_fpca. THis commit fixes this issue

Signed-off-by: Ryan Lerch <rlerch@redhat.com>